packages/calibration/pyproject.toml
[project]
name = "f1-calibration"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "f1-core",
    "pymc>=5.18,<6",
    "numpyro>=0.16,<1",
    "jax>=0.4,<1",
    "arviz>=0.20,<1",
    "netcdf4>=1.7,<2",
    "scikit-learn>=1.5,<2",
    "typer>=0.24,<1",
    "rich>=13",
]

[project.scripts]
f1-calibrate = "f1_calibration.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/f1_calibration"]

[tool.uv.sources]
f1-core = { workspace = true }
